Support Forked PRs
Our options are a bit limited here, but what we can do is open a
restyled PR against the base branch in the upstream repo. It will have
the contributors changes plus style fixups. We can give them
instructions to incorporate those changes locally and update their PR,
or they can leave it up to the maintainer (presumably our customer) to
deal with it however they like.

Closes #4.