Extract restyler-core
Extract a separate binary for *just* the operation of restyling. This
will allow us to focus on making that process robust with proper logging
and understandable behavior, distinct from the logic of clones and
pushes.

This should also make testing easier through something like cram. It
should be feasible to have a test-case scaffold that accepts some
example code to restyle, the assertion can then be set via cram -i.