Fix bug in Fork comment body
We were checking if the Restyled PR was a fork (it never is). Seems less
err-prone to put the conditional front-and-center. As part of this,
totally separate the contents -- even though we duplicate a bit, it's
clearer to do that, than obfuscate things through conditionals. The
latter contributed to this bug, IMO.