Support a --fake option
It became useful to locally test some of the initialization behavior
(specifically RemoteFile) that doesn't require fetching or cloning an
actual PullRequest. Being able to skip that part, exercise the logic we
were interested in, then abort whenever a real Pull Request was needed
was therefore valuable.

So that's what --fake does. Yay laziness.