Handle the multi-path invocation at this level
This will keep restyler images even simpler, since a lot of tools only
accept a single path and we've been working around it using in-restyler
shims so far. This way, we just invoke it multiple times here -- this is
less efficient but we still only do it when we need to, and it's
probably fine anyway. Ultimately, it's easier to understand and easier
on restyler authors.