Shift always-required arguments to command
It's too easy to want to override some arguments and unintentionally
remove a sub-command or inplace flag that must be present for things to
work correctly. In the best case you'll get a confusing error. In the
worst case, the restyler won't run inplace and it'll "find no
differences".

By shifting these things to the non-configurable rCommand attribute, we
ensure they're always there.

NOTE: It's now possible user-configuration is more likely to lead to
passing the same flag multiple times, so we toss a nub on there.