Fix parameter when searching for Restyled PR
If you don't pass head=user:branch (e.g. we were just passing
head=branch) the parameter is completely ignored. This results in us
thinking the first open PR against the base branch is the Restyled PR.

That's clearly bad, but isn't *usually* a problem since its unlikely
we'd ever have a PR with the original PR as its base, if not for
restyling purposes... Except in the case of a forked PR, where the base
of the Restyled PR would be master. In that case, we end up finding the
original PR and closing it! Yikes!