Always force-push to -restyled branches
Fixes restyled-io/restyled.io#71.

There are plenty of ways we could end up with a -restyled branch
existing already, which means we'd need to force-push it. Relying only
on PR presence to know which path to take in this regard is not enough.
If we assume that a branch ending is -restyled is ours --which I think
is a safe-enough assumption for now-- force-pushing the current restyled
content doesn't hurt.