Add a layer of validation to configuration
For objects that we parse in configuration, pre-validate that the keys
of the raw object are all known -- error if not. And include
edit-distance-based suggestions in the error message.

This came from a case where "includes" (not "include") was used in a
configuration, which was silently ignored. The experience was not great.
With this fix, there would be a validation error that has "did you mean
"include"?" in the message -- pretty nice.

Unexpectedly, this was re-usable in namedRestyler too!

This does come with two downsides:

- We can't use Generic anymore

  A minor annoyance, that I'm accepting.

- We have to duplicate the expected keys and the parsed keys

  Not a big deal (IMO): the two expressions will always be a line apart,
  so they're unlikely to drift.