Support disabling comments
For some, the PR statuses are enough but going auto:true would be too
implicit. This allows them to configure to a status-only work-flow.

For now, the default behavior is the same, but we may consider removing
comments in the future. To achieve that, we would probably need a way to
record what kind of configuration users are opting for and then
selecting the most popular behavior as the default.