Add whenConfig helpers
Simplifies the common operation of guarding logic on a configuration
key.

NOTE: One might reasonable expect these to accept (Config -> Bool)
arguments, but doing the env type variable lets us incorporate appConfig
in the asked function without knowledge of the App type within the
Config module.

The callers are currently adding NonEmpty themselves, but we may extract
a whenConfigNonEmpty to handle that, since it's itself common.