Configuration point for requesting review
Adds the following top-level configuration key:

    request_review:
      origin: author|owner
      forked: author|owner

    # short form, for specifying the given value as both at once
    request_review: author|owner

If defined (default is not), the Restyle PR will have a review requested
from either the author or the owner of the original PR, based on the
configured value(s).

Author makes sense for the (same-)origin case because the author is who
would incorporate the style fixes into their PR. Owner makes sense for
the forked case because the Restyle PR includes the original
contribution and any style fixes as one complete PR, to be reviewed by
the project owner. Therefore, these are the defaults.

Both keys are optional in the long form, meaning

    request_review: {}

Is a valid configuration that enables the feature with default behavior.