Revert replacing Circle with Buildkite
I've had a hell of a time keeping my own buildbox stable. Despite making
it an uber-expensive c5 with a large disk, I face constant scaling
issues with any minor workload. Sorry Buildkite.

Even if I could get this to work, they're not really geared towards my
Docker CI workflow: lots of brittle and/or non-obvious stuff going on.

This reverts commits:

- ebafdeb01fb9a49fc05fc8253a3ef0481ce5d32c.
- e20f2aef2d1c00dcab3777d0724a02839d422285.
- a1dbc1eafe70b43ba6c8aa54f7ede6ece4e41f1c.
- dc9826ab7986d279952ae6433bf79280ce666065.
- 2e1d679582e6068040dc476d194253636ede15ad.

My new plan is to run a "buildbox" just for docker operations that still
happen from on Circle, as a replacement for Quay.io and Clay.run.