Handle attempts to read invalid files
The interpreter check is best-effort, we should make sure we don't throw
exceptions if we encounter unreadable files. This is important because
this change happens even if the file is excluded.

That second part might be surprising in general. We would have to
separate the notion of "include" vs "don't exclude" in our logic to
adequately handle it.