Slice HasGit out of HasProcess
Most git actual require, not just the executing a process, but also
being in a valid clone. Not having this distinction let us to attempt to
delete our restyled branch (via git push) from within a setup routine
that had not yet cloned.

By separating our those functions in a new Has class, and tagging
actions appropriately, we would not be able to use the git-push approach
from within setup, and the types now force us to correct this.