Add oHostDirectory to Options
We need this to support running restyle(r|-path) itself via Docker
locally. We added a --host-directory option to restyler and look for a
HOST_DIRECTORY variable in restyle-path. This is primarily for
restyle-path, as running restyler locally isn't particularly useful
(beyond, perhaps, testing some local development). But both are now
possible.

Production was working without this because everything happens in /tmp,
which is mounted into all containers. This means the container and host
paths match, so the restyler->some-restyler --volumes are implicitly
correct.

When running locally, the host directory would be something like
"/home/me/code/foo", which is not known for the restyler->some-restyler
--volume, unless we specify it in the invocation -- which this option
makes possible.

This required introducing a lot of error fields in the Options values
used in simpler environments (restyle-path and tests). This is a strong
indication we should make finer-grained classes so that the ultimate
chain of functions that needs this could use (e.g.) HasHostDirectory
instead of the big HasOptions class. This has been deferred for now as
those error fields are not present in any paths through the main
restyler executable.