Avoid creating an empty PR
Fixes #57.

For example, Brittany and Stylish-Haskell both format import statements,
but handle hiding differently: Brittany will move the hiding and Stylish
will move it back.

Before this change, we'd see each of those commits as having made
changes and decide to proceed. This would result in a PR with commits
but no differences -- a confusing experience. Now we check for overall
change-paths before proceeding.