ignore_labels
When we process a PR, if its labels intersect the configured `ignore_labels` key
(which is `restyled-ignore` by default), ignore that PR.

This provides an explicit way to tell Restyled to stop processing a PR, without
introducing the confusion of implicit behavior in the solution offered by #92.

One possible surprise is that if you cause this behavior mid-PR (by adding the
label, or adjusting configuration) any existing Restyled PR will not be cleaned
up. This is the same as any form of mid-PR disable-ment.