Implement global exclude configuration
We have trouble with projects that have committed dependency files that
are not actually part of their project. These files tend to be numerous
and exotic, which both slows us down and causes errors, for no benefit.

This patch implements a global exclude configuration and defaults it to
the handful of directories we see that cause this trouble.

NOTE: Restyler include happens after this and cannot pull files back in;
it's as if they don't exist. Users for whom this doesn't work must
disable (set to empty) or adjust the exclude key itself.