Pump the breaks on object-style Config
This removes the "Exclude" feature for now.

It seems more useful to be able to repeat Restylers to run more than once at
different overrides (see #51). The two styles are fundamentally opposed, so
we'll go back to the way it was to explore the other, for now.

Users most likely never moved off the list-style anyway, so this blip is unlikely
to be noticed. If anything, this should make a few edge-cases better for existing
users.