Test and fix const argument-order bug
I (incorrectly) assumed insertWith accepted the present key first, which
meant (insertWith const) would only insert a key if not already present.
That appears to be backward, so (insertWith const) is just insert,
always overriding any already-present key.

This meant that our logic in upgradeEnabled was, instead of doing
nothing for already-present enabled keys, always setting the enabled key
based on that hardcoded disabledRestylers list. Since that list is only
as up to date as when enabled was introduced, newer disabled-by-default
Restylers were being enabled-by-default instead. I believe that is only
ormolu so far.

Anyway, this commit:

- Exposes the bug through a test

- Extracts a name for the easy-to-get-wrong functions

  It's used twice, but in the second use-case I believe the object being
  inserted into is always known to not have the colliding key.

- Re-uses the extracted function, passing the tests