Support lifting Restyler container restrictions
Set UNRESTRICTED=1 to remove all container restrictions on the
Restylers, e.g. net-none, cap-drop-all, cpu-shares, and memory.

This is specifically to avoid having cap-drop-all when run locally via
restyle-path, because that causes the root-run Restylers to not have
access to user-owned local files (they're root-owned on production, so
it all still works there). We could introduce a variable to specifically
affect cap-drop, but this bigger, simpler hammer seems fine enough.