Set UNRESTRICTED=1 in restyle-path
The restylers are run as root and the local files are owned by the local
user. With cap-drop=all even root can't (re)write these files. On
production everything is root-owned, so it's fine, but locally we need
to lift the cap-drop.

Since we're in a rarely-used, local context with restyle-path, we
consider any risk associated with this to be acceptable.