Don't close a Restyle PR unless it's Open
With the previous re-open feature, we started loading up the Restyle PR
even if it's Closed, this means we see it (and attempt to close it) at
then end of a no-differences process. This won't work in the case that
it's been merged and the head branch auto-deleted automatically, because
our attempt to delete that reference (again) is an error.

The fix is to simply avoid trying to close a closed PR. And if we avoid
that, we won't also attempt to delete the reference.