Revert "Find existing Restyle PRs even if Closed"
This reverts commit c3b21020e034d9b0a845fc88c130f41b38718d54.

We see constant 422 Unprocessible Entity error with GitHub claiming we
can't re-open the Restyle PR. They aren't giving a reason, but I suspect
it has to do with the PR head branch being deleted and re-created.

Either way, this workflow will require more complexity to function, so
it's being reverted for now until more time can be spent on it.