Configurability of Restyled branch_name
Closes #85.

The hard-coded pattern of {base}-restyled for Restyled Ref names doesn't
always work, especially when other bots' PRs expect names like
{namespace}/* to be branches they can freely edit: a branch named
other-bot/foo-restyled can cause problems.

Therefore, we want to give flexibility to users to work around this via
their configuration:

    branch_name: "restyled/{base}"

Would result in restyled/other-bot/foo, which will not be problematic to
the bot that thinks other-bot/-prefixed branches are their own.

The most major change here was dealing with the case where:

1. The Restyle PR was opened,
2. The user ignored it
3. The original PR is closed,
4. Restyler fails fetching or checking out the head branch

This is not an edge case since closing a PR can very often mean deleting
its head branch immediately (GitHub can now helpfully automate this),
which will cause such a failure on Restyle.

In such a case, we're obviously unable to load the head branch's
configuration, but that now means we can't find the branch_name, which
means we can't find the Restyled PR, which prevents cleaning up the PR
we created at (1).

The approach in this commit is to fall back to the default branch's
configuration when we can (i.e. when the clone succeeds, but fetches or
checkouts fail). As long as it's using the same branch_name as the
just-closed PR, things will work the same way. If it's not, we will skip
the clean up, or --absolute worst-case-- we would clean up a different
PR. That final bit seems bad, but the actual negative outcome of such a
misfire should be minimal and (hopefully) it rarely happens, so we
accept it as an un-handled edge-case.