Move configuration documentation to wiki
I see a fair amount of configuration errors that come from users
uncommenting example blocks in a naive way, which breaks indentation.

Moving all the prose comments to the wiki, leaving only YAML in the
source should make it easier to see the overall structure and
indentation of the source file itself. And copy/paste from wiki code
blocks is more likely to go in correctly at top-level and not break
things. GitHub markdown also allows some better formatting.

The trade-off is now things need to be kept in sync, so we'll see.