Deprecate restylers key, add overrides key
The current restylers key has the annoying behavior that if you use it
to modify one Restyler, you no longer run any of the other default
Restylers (unless you also add them explicitly).

Changing this would impact current users too drastically, so this commit
adds a new key which is basically the same, except corrects this
behavior.

The naming is somewhat important: "restylers" as a name does signal that
you're defining the list of Restylers to run, full stop. While
"overrides" indicates you just want to override some thing(s) and not
lose others.

We deprecate the restylers key as part of this, which means that users
who indeed wish to configure only a specific set of Restylers to run
must be able to do so, even as Restylers are added to the system
defaults later, without having to update their configs in response.

To support this, we accept the override "name" key as a Glob. This means
that such users would use the following configuration:

    overrides:
      - one
      - two
      - "*":
          enabled: false

The initial items declare what to run, and the final wildcard ensures no
other (not yet known) Restylers are ever run. The single wildcard will
be the most common usage, but it is a glob, so configuring (e.g.)
"prettier*" at once, say to use a newer image, is also supported.

The behavior of interpreting a top-level list as the restylers key
remains for now. When we fully remove the Restylers key we will also
remove this behavior and force everyone to always configure an object at
top-level, and so choose the overrides key. Maybe we'll bring back the
list convenience as a short-cut for overrides after that.

Closes #197.