Process interpreters before include
Fixes #104.

Before, we would send each path through include processing first, and if
it wasn't included, we then check interpreter. This had a few downsides,

1. For every un-included path, we would read the file for interpreter
   information repeatedly for every Restyler that had such configuration

2. If you explicitly exclude a file, it would still (and in fact,
   consequentially) be checked for interpreter and restyled if matched

The new logic solves that by,

- Reading any (executable) paths once, and attaching interpreter
  information to the list of paths before include-processing

- Checking interpreters first and, if matched, adding an artificial
  include for that file before doing the typical include filtering

In other words, a config such as,

    restylers:
      - shfmt:
          include:
            - "**/*.sh"
          interpreters:
            - sh

will behave like one might imagine,

    restylers:
      - shfmt:
          include:
            - "**/*[interpreter==sh]"
            - "**/*.sh"

would behave.

This means that including a file based on interpreter (and not pattern)
behaves according to the intuition that it's included (first) and then
possible to exclude later,

    restylers:
      - shfmt:
          include:
            - "**/*[interpreter==sh]"
            - "**/*.sh"
            - "!bin/bad-script"