Use typical JSON parser for Restyler information
This parser is not run on user-authored data, it's run on our
restylers.yaml manifests. Therefore, the careful handling of extra keys
and "Did you mean?" errors is not valuable.

In fact, the strict parser (errors on extra keys) makes it harder to
evolve that Yaml file. For example, we'd like to start including the
metadata key so tooling can operate using only the manifest and not need
to go back to info.yaml files for (e.g.) tests and languages
information.