Ignore symbolic links
Symlinks aren't anything useful for us. If they point at a file in the
repository, and if that file needs Restyling, we will see it directly.
If they point outside the repository or to themselves it can cause us a
lot of trouble.

Due to the symmetry with isFileExecutable, we chose to make our own name
and not call this isPathSymbolicLink to match System.Directory.