Use Statsd.Dummy if statsd options not given
This shouldn't be necessary.

Statsd is a UDP protocol and sending stats to a host without a server
running should be perfectly fine. However, I've found that when running
locally, the completion of the withStatsClient block raises an hFlush
"Connection refused" exception.

To mitigate this, we make the host/port optional (instead of just
defaulted). If neither are given, we use the Statsd.Dummy client. If any
are present, we use a real client with defaulting.