Find Restyled PRs by head only
Given a PR with head pb/fix, we'd locate a Restyled PR by looking for
one where:

    base == pb/fix &&
      head == restyled/pb/fix

This is a very precise search but fails if GitHub has automatically
switches the base after pb/fix was merged, as it so helpfully does.

We could update this to,

    (base == pb/fix || base == {default branch}) &&
      head == restyled/pb/fix

But that seems complicated, and would miss a case of a Restyling on a
branch-from-branch situation.

We could simplify to,

    head == restyled/pb/fix

But false-positives are scary, considering they could result in a
force-push over a user's unrelated Pull Request.

Therefore, we'll do

    head == restyled/pb/fix &&
      user =~ ^restyled-io

Having the check on user will prevent force-pushes onto users no matter
how well this works (or not).

Fixes #150.