Simplify clone and changed-files logic
Trying to figure out changed-files in a PR is fraught with land-mines.

The only way to reliably do it in a local checkout (by finding, and
diffing against, the merge-base), is to have a complete clone.

Follow me on my attempts to solve that:

You would think you could take the base.sha in the PR event as given,
fetch only as deep (on the base branch) as you need to find that, then
diff against it. This works reliably in a same-origin PRs, but in a Fork
I get incorrect results.

    -- A -- B -- C -- D
            \
             \-- E -- F

The ideal merge-base is B, and same-origin PRs seem to reliably include
that as base.sh. I don't know what base.sha is for the Fork case, but it
doesn't work. Maybe D?

Cool.

You would also think you could fetch only as deep (along both branches)
as you need to find a merge-base, then diff against it. This works
reliably in a Fork, but in a same-origin PR I get incorrect results.

Cool.

This is likely due to git finding the "best" merge-base, meaning it can
succeed on a sha even though there is a better one (the right one)
further back. We have no way, externally, to know to keep looking.

Luckily, there is an easier way: we can just ask GitHub what the changed
files are via the pull/N/files endpoint. :facepalm:

Not having to work out changed files locally means:

- We have no reason to fetch the base at all
- We have no reason to fetch any further than --depth=1

So that's what gitCloneBranchByRef does. It is functionally equivalent
to a "shallow clone":

    git clone --branch x --depth 1

But spelled out as init/remote-add/fetch so that it will work with the
special GitHub pull/N/head ref. This is the only way to do it for a
Fork, and it works fine for same-origin too. Yay for a single code-path.

As part of this, I in-lined all the non-HasGit functions to minimize
what's going on outside the typeclass. Moving gitCheckout into the
type-class meant it had to happen after restylerSetup, but placing it
just alongside the push(es) is much clearer -- particularly, the use of
pullRequest vs restyledPullRequest functions for the restyled branch
name is much less surprising now that one happens in the create path and
the other the update path.