Fix invalid usage handling in restyle-path
Without a leading ":" in the getopts format string, you are not expected
to handle an unknown option (\?) vs a missing argument (:) yourself, so
my handling of those case was unnecessary and broken. Both paths print
their own (perfectly acceptable) error messages and go through the \?
case, which we can simplify.