Install jo in Restyler image
The restyled/agent also uses the Restyler image to output "skipped"
messages for (e.g.) Restyled PRs. These messages come out as plain text
which is interpreted as DEBUG and not visible to users immediately. `jo`
is a great way to output this as a properly-structured log.

It's important to use `jo` and not just build up the string of JSON,
since we use this for other error messages, which can have arbitrary
content and produce invalid JSON if handled naively.