Only fetch Restyled PR when necessary
This is only needed if we're closing, opening, or updating the Restyled
PR, which is very localized. Keeping it out of App is simpler. Also,
fetching it early can exacerbate race-conditions when two Restyler runs
attempt to create the same Restyled PR because it didn't exist when the
both ran the earlier fetch.