Add config.also_exclude
The current `exclude` could (and can still) be used to fully set what we
exclude, but in order to simply add a new pattern, you have to know,
re-specify, and forever maintain, our default choices in your own
config. This is annoying and error-prone.

The new key supports simply adding a pattern while continuing to respect
ours, even if they change in the future.

Closes #177.