Handle skip logic within-CLI
Rather than having some validations run in our agent, and others run in
the CLI, let's have them all run in the CLI. The Agent can pass along
the `restylerEnv`, keeping ownership over things like disabling and plan
restrictions in the API, where it belongs.

Originally, I wanted it to be efficient when things are skipped, by not
running the CLI at all for such cases. But we have been, since we use
the same image to emit the skipped logging details. This is no
different.

By doing it in CLI, we can decide if/how PR statuses are updated in
response to skipping. For example, a plan restriction is actually an
error and emitting the PR status is a chance for users to be aware and
potentially upgrade their plans.