Tweak text-fixture Restylers
The primary motivation for this is to make a realistic `rImage`, which
we're about to work on and will need to test. This required passing name
to `someRestylers`, which changed various things. While changing those
areas of `ConfigSpec`, I spotted a new useful `assertLoadsRestyler`
helper, which I refactored to generally.