Optionally parse individual image parts
Currently, users have to specify a full image or not,

```yaml
image: restyled/restyler-foo:v1.0.0
```

This means that to use an alternative tag, they have to get the registry
and name right.

This patch makes it possible to specify an object of fields,

```yaml
image:
  registry: restyled
  name: restyler-foo
  tag: v1.0.0
```

All fields are optional and will default from the manifest image, which
means users can simply override `tag` and not worry about the rest,

```yaml
image:
  tag: v1.0.0
```

This will be useful when we start shipping "series" tags and users will
be interacting with this point of configuration more.

```yaml
image:
  tag: v1
```