Address ENV parsing errors on --help
Since `restyle-gha` has required ENV (`GITHUB_TOKEN`), it was failing
when we run `--help`. We can correct this by parsing ENV after options.

Additionally, we can extract the help from the ENV parser and add it as
a footer to `--help`. That help doc isn't great, but it's better than
nothing for now, and we may upstream improvements.