Fix image-cleanup in one-at-a-time case
Before we would clean up images as part of run, so if we were running
the same restyler multiple times (once per file) we would remove and
then immediately re-pull (as part of `run`) on each file.

This refactor moves it out to a more reasonable place. Along with the
rename, the code looks to more clearly represent the 3-step process of
pull, run, remove.