Fix exclude handling with explicit arguments
This logic would only apply our excludes to the arguments given, not the
files post-expanded. This means if you call `restyle .` it effectively
ignores excludes.

This was mostly fine as usage on a PR will always pass fully expanded
paths, but breaks locally, particularly in our integration tests where a
`.git` directory would be picked up by shebang.

The fix is to apply excludes after expansion, but we can also apply them
before as an optimization (no use expanding if excludes happen to hit at
that level already).