Avoid hidden files in expanding arguments
This doesn't change anything when paths are given explicitly (such as in
the list of changed files on PRs), but when directories are given in
local or test usage, and we are expanding them, we now exclude hidden
files. I think this makes the most sense: match the behavior of a
recursive `ls`.