Prevent empty changesets when restylers offset each other
Sometimes, two restylers can "fight". One changes code one way and then
another puts it exactly back. When this happens, each restyler will
generate their own commit, but the overall changeset will be empty.

If we don't explicitly check for this, using the create-pull-request
action will still produce a PR because there are commits, even if the
overall diff is empty. This empty PR might just be merged out of habit
or process, triggering the process to repeat again.

So we'll explicitly check:

- If we made changes,
- If those changes made commits
- If there's no diff against our first commit's parent (`{sha}^`)
- Then we'll do a `git reset --hard` back to it