Ignore git-clean failures
Some restylers create files that we do not have permission to clean up.
I might've expected that to always be the case, since they run in a
container as `root`, but many restylers create files we can and do clean
up. Attempting it and ignoring permissions errors seems fine; it was
best-effort anyway.