Rewrite option, env, and config handling
This completely replaces our options, environment, and configuration
handling with the off-the-shelf library `opt-env-conf`.

All settings are now normalized and parsed through the three sources, we
have a much richer help and man-page output, and we dropped a massive
amount of code.

It's hard to say what users were relying on, so we aimed to be somewhat
compatible, but not completely. Options and configuration are backwards
completely compatible because we know users rely on the latter and
`actions/run` uses the former. We gave ourselves more liberties with
environment variables, in the interest of normalization and simplicity.

Here is a possibly-not-exhaustive summary of changes:

* Options (backwards-compatible)
  * Added `--config-file`
  * Added `--{also,}exclude`
  * Added `--ignore-{authors,branches,labels}`
  * Added `--restylers-version`
  * Added `--host-directory`
  * Added `--restyler-{restricted,net-none,cpu-shares,memory}`
  * Every switch has both `--x` and `--no-x`.
* Environment (backwards-compatible)
  * Added `CONFIG_FILE`
  * Added `FAIL_ON_DIFFERENCES`
  * Added `{ALSO_,}EXCLUDE`
  * Added `IGNORE_{AUTHORS,BRANCHES,LABELS}`
  * Added `RESTYLERS_VERSION`
* Environment (**not** backwards-compatible)
  * Changed `NO_IMAGE_CLEANUP` to `IMAGE_CLEANUP=False`
  * Changed `NO_PULL` to `PULL=False`
  * Changed `NO_COMMIT` to `COMMIT=False`
  * Changed `NO_CLEAN` to `CLEAN=False`
  * Changed `UNRESTRICTED` to `RESTYLER_RESTRICTED=False`
* Config (backwards-compatible)
  * Added `logging.*`
  * Added `dry_run`
  * Added `fail_on_differences`
  * Added `restylers_manifest`
  * Added `docker.*`
  * Added `git.*`
  * Added `ignore.*` (`ignore_*` is still supported)