Have some compatibility with Blammo settings
Because `LOG_*` variables are parsed outside and before we parse our
`opt-env-conf` settings, they are always overridden by the fact that we
use the config file to specify our defaults.

If we also parse those variables ourselves, we can ensure they are
respected on top of our defaults if given.

This works perfectly with `LOG_COLOR` since we parse the same structure
and can just read it at its name, but `LOG_LEVEL` has a caveat: it will
be respected when set, no matter what. That's because we are parsing
boolean switches so we can't fold that into parsing the levels value at
the same time and we must do one or the other via `(<|>)`.

In other words,

```console
LOG_COLOR=always restyle --color=never
```

Works as you expect, but

```console
LOG_LEVEL=debug restyle --trace
```

Does not.

I think this is acceptable.