Filter in versions instead of filtering out RCs
Since we're not using a proper JSON parser, we can't easily filter by
prerelease or draft attributes, instead we need to filter on the tag
values themselves. Filtering *out* `-rc$` was effective when those were
the only kind of prereleases we made, but as we experiment with Semantic
Release, this changes. It's better to use positive filtering anyway.