feat: format log messages differently
This format is inspired by Renovate's logs and is a better way to
represent metadata for CLIs than the key-values of the default `tty`
format. We also don't need timestamps as GitHub Actions can add them
itself.

This does mean `LOG_FORMAT` is no longer respected; we always use this
format.