feat: auto-enable restylers based on groups and priorities
Rather than having a fixed default of enabling or disabling various
restylers, support a new `auto_enable` key which allows specifying:

- Patterns for configuration files used by this restylers
- An optional group (name and priority), to declare restylers of which
  only one should be enabled at a time

Given this, we can:

- Check if any configuration patterns match
- Auto-enable any restylers where a configuration file is present
- If no configuration paths exist, auto-enable the first one in the
  group, by priority, **unless** the user has explicitly enabled one in
  that group already

This means, for example,

Between Fourmolu, Brittany, and Stylish Haskell, we enable Brittany by
default. But if the user has a `.fourmolu.yaml`, we'll enable (only)
Fourmolu instead.

Between jq and prettier-json, we enable prettier-json by default. But if
the user enables jq, we won't.

This is behavior neutral for now, while our restylers manifest only uses
`enabled: true|false` keys. But next we can start creating groups and
replacing that with `auto_enable`.