fix!: flip defaults for commit and clean options
Committing and cleaning makes sense when run as an action, and don't
make sense when run locally. Therefore, we should make you supply the
positive options, since that's encoded in the action. The option-less
behaviors should be those desired locally.

It's possible committing also makes sense as the locale default, but I'm
not sure.